BankAccount - polish

+++ week7/day2/week7day2/src/main/java/com/greenfoxacademy/Model/BankAccount.java
@@ -8,6 +8,6 @@ import lombok.Getter;
 public class BankAccount {
 
   String name;
-  float balance;
+  int balance;
   String animalType;
 }
